FROMLIST: scsi: ufs: Fix ufshcd_hold() caused scheduling while atomic
The async version of ufshcd_hold(async == true), which is only called
in queuecommand path as for now, is expected to work in atomic context,
thus it should not sleep or schedule out. When it runs into the condition
that clocks are ON but link is still in hibern8 state, it should bail out
without flushing the clock ungate work.

Bug: 143632303
Change-Id: I14d200901274c2e630bf0205445c1f9c7e053f95
Link: https://lore.kernel.org/patchwork/patch/1149893/
Signed-off-by: Can Guo <cang@codeaurora.org>
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: Matthias Maennich <maennich@google.com>