ANDROID: gki_defconfig: Set IKHEADERS back to =m
This partially reverts I56e59e1a6232dada090d720f43e0d1036f43d064.

The build/runtime vintf check validates that CONFIG_IKHEADERS=m, it is
not allowed to be set to =y. While we figure out what the right course
of action is here, modify it back to =m.

Bug: 139431025
Test: TreeHugger
Change-Id: I3624fa4eb40a8aa726275027aa2b2d5bd635ceda
Signed-off-by: Alistair Delva <adelva@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>